fix(matting): find it — WebGPU Conv is wrong when C_in %3==0 and %4!=0
Root cause, after every whole-model hypothesis had been eliminated.

ONNX Runtime's WebGPU provider returns wrong values from Conv when the input
channel count is divisible by 3 and not by 4. Minimal reproducer: one dense
3x3 Conv, 16 output channels, 16x16 spatial, varying only C_in. Across a
1..176 sweep the failing set is exactly

    6, 9, 15, 18, 21, 27, 30, 33, 39, 171

at 17-20% off in L1. Every multiple of 4 is exact. C_in=3 is special-cased
upstream and works. Identical on ORT 1.24.2 and 1.27.0.

RVM walks into this exactly once. Its decoder concatenates the 3-channel
source image onto its skip connections, so Conv_200 has W=[80,171,3,3]. One
conv out of 353 nodes, and the whole matte is destroyed.

Found by sub-model extraction, comparing each probe across providers: every
tensor matched to ~1e-7 up to and including Concat_199 — the failing conv's
own input — and then Conv_200's output was 44% off. A caveat that cost a
pass and is now written down: compare with a RELATIVE tolerance of about
1e-3, because fp reassociation alone moves an L1 sum by ~1e-7 and exact
equality marks every probe divergent, localising nothing.

The fix is an exact identity: pad the input channels to the next multiple of
4 with an ONNX Pad, and zero-pad the weight tensor along C_in to match. The
appended weights are zero, so the appended channels contribute nothing
whatever they hold; only the kernel path changes. Verified end to end on the
same frame, 512x288, 10 passes:

    stock  webgpu   fg>0.5 0.000  mean 0.002  max 0.185   6.71 ms/frame
    padded webgpu   fg>0.5 0.227  mean 0.227  max 1.000   7.34 ms/frame
    padded cpu      fg>0.5 0.227  mean 0.227  max 1.000  41.06 ms/frame

Padded-WebGPU matches CPU exactly, which is what proves the rewrite is an
identity, and it is 5.6x faster than the CPU provider. So GPU matting needs
no CUDA and costs no vendor neutrality.

Not yet wired into the daemon: the padding is currently a Python script
against the .onnx. Doing it at load time in Rust needs onnx protobuf
round-tripping (prost is already in the tree), because prost drops unknown
fields and a partial schema would silently corrupt the model.

Adds examples/ep_diff.rs, a generic two-provider comparator — matte_sweep
hardcodes RVM's six inputs and so could never have answered "is this one
operator wrong?".

+++ crates/cleanroom-matting/examples/matte_sweep.rs
@@ -150,21 +150,21 @@ fn main() {
     let preseed = env("CR_PRESEED").is_some();
     let mut state: Vec<(Vec<i64>, Vec<f32>)> = if preseed {
         // RVM-mobilenetv3's recurrent channels, at 1/2, 1/4, 1/8 and 1/16 of the input.
-        [(16, 2), (20, 4), (40, 8), (64, 16)]
+        [(16i64, 2i64), (20, 4), (40, 8), (64, 16)]
             .iter()
             .map(|(c, div)| {
                 let (rh, rw) = ((h as i64) / div, (w as i64) / div);
-                (
-                    vec![1, *c as i64, rh, rw],
-                    vec![0.0f32; (c * (rh * rw) as i64) as usize],
-                )
+                (vec![1, *c, rh, rw], vec![0.0f32; (c * rh * rw) as usize])
             })
             .collect()
     } else {
         (0..4).map(|_| (vec![1, 1, 1, 1], vec![0.0f32])).collect()
     };
     if env("CR_SHAPES").is_some() {
-        eprintln!("[s] seed shapes {:?}", state.iter().map(|s| &s.0).collect::<Vec<_>>());
+        eprintln!(
+            "[s] seed shapes {:?}",
+            state.iter().map(|s| &s.0).collect::<Vec<_>>()
+        );
     }
 
     // Optional 6th argument: how many frames to run. The default settles the recurrent
